requirements.txt
requests
geocoder
python-dateutil
colorama
ics
